requirements_app.txt
streamlit>=1.26.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=9.5.0
torch>=2.0.0
pyyaml>=6.0.0
pandas>=2.0.0
